Actualizado Roles Admin, Vendedor, Comprador. Agregado funcionalidades completas para cada rol, actualizado inferfaz de Usuario en todos los roles, conestado base de datos de Supabase y añadido base de datos estructurada para manejo de la informacion en la app.

+++ app/src/main/java/com/example/cakebyteapp/domain/usecase/GetProductByIdUseCase.kt
@@ -8,5 +8,5 @@ import javax.inject.Inject
 class GetProductByIdUseCase @Inject constructor(
     private val repository: ProductRepository
 ) {
-    operator fun invoke(id: Int): Flow<ProductEntity?> = repository.getProductById(id)
+    operator fun invoke(id: Int): Flow<ProductEntity?> = repository.getProductById(id.toLong())
 }
